Move AI chat into shared templates
