+++ css/cabinet.css
@@ -100,279 +100,6 @@
     background-color: rgba(220, 53, 69, 0.2);
 }
 
-/* AI Chat */
-.ai-chat-toggle,
-.ai-chat-icon-btn {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    width: 40px;
-    height: 40px;
-    border: 1px solid var(--border-color);
-    border-radius: 8px;
-    background-color: var(--bg-secondary);
-    color: var(--text-primary);
-    cursor: pointer;
-    font-size: 1.05rem;
-    transition: background-color 0.2s ease, color 0.2s ease, border-color 0.2s ease;
-}
-
-.ai-chat-toggle:hover,
-.ai-chat-icon-btn:hover {
-    background-color: var(--card-bg);
-    color: var(--button-primary);
-    border-color: var(--button-primary);
-}
-
-.ai-chat-backdrop {
-    position: fixed;
-    inset: 0;
-    background: rgba(17, 24, 39, 0.36);
-    z-index: 500;
-}
-
-.ai-chat-panel {
-    position: fixed;
-    top: 0;
-    right: 0;
-    z-index: 501;
-    display: flex;
-    flex-direction: column;
-    width: min(460px, 100vw);
-    height: 100vh;
-    background-color: var(--card-bg);
-    border-left: 1px solid var(--border-color);
-    box-shadow: -8px 0 24px rgba(15, 23, 42, 0.16);
-    transform: translateX(100%);
-    visibility: hidden;
-    pointer-events: none;
-    transition: transform 0.22s ease, visibility 0.22s ease;
-}
-
-.ai-chat-panel.open {
-    transform: translateX(0);
-    visibility: visible;
-    pointer-events: auto;
-}
-
-.ai-chat-header {
-    display: flex;
-    align-items: flex-start;
-    justify-content: space-between;
-    gap: 1rem;
-    padding: 1.1rem 1.25rem;
-    border-bottom: 1px solid var(--border-color);
-    background-color: var(--bg-secondary);
-}
-
-.ai-chat-header h2 {
-    margin: 0 0 0.2rem;
-    font-size: 1.15rem;
-    color: var(--text-primary);
-}
-
-.ai-chat-status {
-    display: inline-flex;
-    align-items: center;
-    min-height: 1.25rem;
-    color: var(--text-secondary);
-    font-size: 0.8rem;
-}
-
-.ai-chat-body {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    padding: 1rem 1.25rem;
-    overflow-y: auto;
-}
-
-.ai-chat-config {
-    flex: 0 0 auto;
-    border: 1px solid var(--border-color);
-    border-radius: 8px;
-    padding: 1rem;
-    background-color: var(--bg-primary);
-}
-
-.ai-chat-config-grid {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 0.75rem;
-}
-
-.ai-chat-config-wide {
-    grid-column: 1 / -1;
-}
-
-.ai-chat-checkbox-label {
-    display: flex;
-    align-items: center;
-    gap: 0.45rem;
-    color: var(--text-primary);
-    font-size: 0.85rem;
-    cursor: pointer;
-}
-
-.ai-chat-checkbox-label input {
-    width: 1rem;
-    height: 1rem;
-    accent-color: var(--button-primary);
-}
-
-.ai-chat-config-actions {
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-    gap: 0.5rem;
-    margin-top: 0.9rem;
-}
-
-.ai-chat-presets {
-    flex: 0 0 auto;
-    display: grid;
-    grid-template-columns: repeat(3, minmax(0, 1fr));
-    gap: 0.5rem;
-}
-
-.ai-preset-btn {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 0.4rem;
-    min-height: 42px;
-    padding: 0.55rem 0.45rem;
-    border: 1px solid var(--border-color);
-    border-radius: 8px;
-    background-color: var(--bg-secondary);
-    color: var(--text-primary);
-    font: inherit;
-    font-size: 0.82rem;
-    cursor: pointer;
-    transition: background-color 0.2s ease, border-color 0.2s ease, color 0.2s ease;
-}
-
-.ai-preset-btn:hover {
-    background-color: var(--card-bg);
-    border-color: var(--button-primary);
-    color: var(--button-primary);
-}
-
-.ai-chat-messages {
-    flex: 0 0 auto;
-    display: flex;
-    flex-direction: column;
-    gap: 0.75rem;
-    min-height: 180px;
-}
-
-.ai-chat-message {
-    max-width: 88%;
-    padding: 0.7rem 0.85rem;
-    border: 1px solid var(--border-color);
-    border-radius: 8px;
-    background-color: var(--bg-secondary);
-    color: var(--text-primary);
-    line-height: 1.45;
-}
-
-.ai-chat-message-user {
-    align-self: flex-end;
-    background-color: rgba(59, 130, 246, 0.12);
-    border-color: rgba(59, 130, 246, 0.26);
-}
-
-.ai-chat-message-assistant {
-    align-self: flex-start;
-}
-
-.ai-chat-message-author {
-    margin-bottom: 0.25rem;
-    color: var(--text-secondary);
-    font-size: 0.74rem;
-    font-weight: 600;
-}
-
-.ai-chat-message-text {
-    font-size: 0.9rem;
-    white-space: pre-wrap;
-}
-
-.ai-command-queue {
-    flex: 0 0 auto;
-    display: flex;
-    flex-direction: column;
-    gap: 0.6rem;
-    border-top: 1px solid var(--border-color);
-    padding-top: 0.9rem;
-}
-
-.ai-command-empty {
-    color: var(--text-secondary);
-    font-size: 0.86rem;
-}
-
-.ai-command-card {
-    border: 1px solid var(--border-color);
-    border-radius: 8px;
-    padding: 0.85rem;
-    background-color: var(--bg-primary);
-}
-
-.ai-command-card-header {
-    display: flex;
-    align-items: flex-start;
-    justify-content: space-between;
-    gap: 0.75rem;
-    margin-bottom: 0.5rem;
-}
-
-.ai-command-title {
-    color: var(--text-primary);
-    font-size: 0.9rem;
-    font-weight: 600;
-}
-
-.ai-command-status {
-    flex-shrink: 0;
-    border-radius: 999px;
-    padding: 0.18rem 0.55rem;
-    background-color: rgba(34, 197, 94, 0.12);
-    color: #16a34a;
-    font-size: 0.72rem;
-    font-weight: 600;
-}
-
-.ai-command-meta {
-    color: var(--text-secondary);
-    font-size: 0.8rem;
-    line-height: 1.4;
-}
-
-.ai-command-actions {
-    display: flex;
-    justify-content: flex-end;
-    gap: 0.45rem;
-    margin-top: 0.7rem;
-}
-
-.ai-chat-composer {
-    display: flex;
-    align-items: flex-end;
-    gap: 0.65rem;
-    padding: 0.9rem 1.25rem 1rem;
-    border-top: 1px solid var(--border-color);
-    background-color: var(--bg-secondary);
-}
-
-.ai-chat-composer textarea {
-    min-height: 74px;
-    max-height: 160px;
-    resize: vertical;
-    font-family: inherit;
-}
-
 /* Cabinet Layout */
 .cabinet-layout {
     display: flex;
@@ -1658,31 +1385,6 @@
 @media (max-width: 600px) {
     .cabinet-content {
         padding: 1rem;
-    }
-
-    .ai-chat-panel {
-        width: 100vw;
-    }
-
-    .ai-chat-header,
-    .ai-chat-body,
-    .ai-chat-composer {
-        padding-left: 1rem;
-        padding-right: 1rem;
-    }
-
-    .ai-chat-config-grid,
-    .ai-chat-presets {
-        grid-template-columns: 1fr;
-    }
-
-    .ai-chat-composer {
-        flex-direction: column;
-        align-items: stretch;
-    }
-
-    .ai-chat-composer .btn-primary {
-        justify-content: center;
     }
 
     .profile-form,
